requirements.txt
pandas
numpy
tqdm
requests
scipy
pyproj
xgboost
python-dotenv